requirements.txt
logging
asyncio
pillow
numpy
aiogram